+++ src/writer/junit.rs
@@ -311,6 +311,8 @@ impl<W: Debug, Out: io::Write> JUnit<W, Out> {
                     })
                     .add_testcase(case);
             }
+            // TODO: Report logs for each `Scenario`.
+            Scenario::Log(_) => {}
         }
     }
 
@@ -393,6 +395,8 @@ impl<W: Debug, Out: io::Write> JUnit<W, Out> {
                     sc.name,
                 );
             }
+            // TODO: Report logs for each `Scenario`.
+            Scenario::Log(_) => unreachable!(),
         };
 
         // We should be passing normalized events here,